refactor: split HTTP/2 server into focused modules
